feat: add support for changing OpenAI model
Added a new command to ask the user if they want to temporarily change the OpenAI model when the maximum number of tokens is reached. The new command uses the `AskToChangeModelCli` interface to create a prompt asking the user if they want to change the model. If the user confirms, the `config.Model` property is updated with the new model and the `handleCmd` function is called again. If the user doesn't want to change the model, a message is displayed indicating that the maximum allowed token for the current model was reached and that the user can try a new model running 'autocommit set --model <model>' or decrease the amount of files being committed.